undo: move unsavedChanges() functionality to Command::isClean()
Currently to check for unsaved changes, we have to query the undo
stack and check the global divelist_changed flag. This is done
on desktop in the unsavedChanges() function.

Move this functionality to Command::isClean() to make it available
from mobile.

Yes, this is a layering violation. However, the long-term goal
is to remove the global flag anyway.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>