mobile/undo: consider undo-stack when checking for changes
The mobile code only used the global divelist_changed flag
to check for changed dives or mark the current state as clean.
However, undo-commands don't set this flag, the reason being
that undo/redo might end up in the unsaved state.

Therefore, use the functions of the undo stack. This is a
further step in unifying mobile and desktop.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>