documentation/mobile-manual: create a merged v3 manual
Unintentionally two versions of the mobile manual were created. This brings all
of Willem's changes into the version for Subsurface-mobile v3, and restores the
last version of the v2 mobile manual.

This commit was manually created from a pull request that was signed off by
Willem, given that he wrote most of the text I am copying that SOB into this
commit and will assign authorship to him.

To avoid further confusion, we now have a v2 and a v3 manual, explicitly named.
The corresponding .html.git files were also updated (reflecting our switch to
asciidoctor).

Signed-off-by: willemferguson <willemferguson@zoology.up.ac.za>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>